Updated project structure to improve build times.
Projects now have a boolean filed 'assetsOnly'. If true, then the builder
will not download them at all when downloading projects.

assetsOnly is determined by whether the project has an xrengine.config.ts file.
template-project now does not have this file, so that by default new projects
made in the studio are assumed to be asset-only. For reference, there is a new
template-code-project that has that file.

Projects' thumbnail links are now in package.json -> etherealEngine.thumbnail.
This is so that they can have this thumbnail link without needing an
xrengine.config.ts file.

Builder now ignores asset files via .dockerignore files when doing Docker files,
as assets do not need to be included on the backend and do not need to be built
into front-end minified code, since they are referenced by exact paths separate
from the client-side code.

+++ packages/server-core/src/route/route/route.test.ts
@@ -72,7 +72,7 @@ const updateXREngineConfigForTest = (projectName: string, customRoute: string) =
   const projectLocalDirectory = path.resolve(projectsRootFolder, projectName)
   const xrEngineConfigFilePath = path.resolve(projectLocalDirectory, 'xrengine.config.ts')
 
-  fs.rmSync(xrEngineConfigFilePath)
+  if (fs.existsSync(xrEngineConfigFilePath)) fs.rmSync(xrEngineConfigFilePath)
   fs.writeFileSync(xrEngineConfigFilePath, testXREngineConfig)
 }
